Rename to acamodels (#52)
* Update README

* Rename to acamodels

* Fix all imports

* Rename to acamodels in test workflow
