+++ acamodels/__init__.py
@@ -7,9 +7,9 @@ __version__ = "0.4.0"
 # Imports
 # -----------------------------------------------------------------------------
 
-from datamodels.aca_base import ACABase
-from datamodels.archive_file import ArchiveFile
-from datamodels.identification import Identification
+from acamodels.aca_base import ACABase
+from acamodels.archive_file import ArchiveFile
+from acamodels.identification import Identification
 
 # -----------------------------------------------------------------------------
 # Public API
