+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "datamodels"
+name = "acamodels"
 version = "0.4.0"
-description = "Datamodels based on pydantic used in Python tools developed at Aarhus Stadsarkiv."
+description = "Datamodels based on pydantic used in Python tools developed at Aarhus City Archives."
 authors = ["Nina Jensen <jnik@aarhus.dk>", "Claus Juhl Knudsen <cjk@aarhus.dk>"]
 maintainers = ["Aarhus Stadsarkiv <stadsarkiv@aarhus.dk>"]
 license = "GPL-3.0"
