+++ tests/test_identification.py
@@ -5,7 +5,7 @@
 from pydantic import ValidationError
 
 import pytest
-from datamodels import Identification
+from acamodels import Identification
 
 # -----------------------------------------------------------------------------
 # Tests
